Connection updates

+++ app/src/main/java/com/sandello/ndscalculator/GetRates.kt
@@ -2,7 +2,6 @@ package com.sandello.ndscalculator
 
 import android.content.Context
 import android.os.StrictMode
-import android.util.Log
 import androidx.room.Room
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
@@ -10,14 +9,16 @@ import kotlinx.coroutines.launch
 import kotlinx.serialization.builtins.list
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonConfiguration
+import java.io.BufferedInputStream
+import java.io.BufferedReader
+import java.io.InputStreamReader
+import java.lang.Exception
 import java.net.HttpURLConnection
-import java.net.SocketTimeoutException
 import java.net.URL
 import java.util.*
 
 
 var listRates: List<Rate>? = null
-var url: String? = null
 
 class GetRates {
     fun main(context: Context) {
@@ -26,24 +27,38 @@ class GetRates {
         StrictMode.setThreadPolicy(policy)
         if (checkConnection(context)) {
             GlobalScope.launch(Dispatchers.IO) {
-//                url = URL("http://jedioleg.asuscomm.com:8000").openConnection().getInputStream().bufferedReader().readText()
-                val url2 = URL("http://jedioleg.asuscomm.com:8000")
-                val con: HttpURLConnection = url2.openConnection() as HttpURLConnection
-                con.connectTimeout = 500
-                Log.d("rates", con.inputStream.bufferedReader().readText().toString())
-                val json = Json(JsonConfiguration.Stable.copy(isLenient = true, ignoreUnknownKeys = true))
-                listRates = json.parse(Rate.serializer().list, url!!)
-                val db = Room.databaseBuilder(
-                        context,
-                        AppDatabase::class.java, "rates"
-                ).allowMainThreadQueries().build()
-                db.rateDao().deleteAll()
-                db.rateDao().insertAllCountries(listRates!!)
-                val prefs = context.getSharedPreferences("val", Context.MODE_PRIVATE)
-                if (prefs?.getString("rate", "") == "") {
-                    val editor = prefs.edit()
-                    editor?.putString("rate", db.rateDao().findByCountry(Locale.getDefault().country).rate.toString())
-                    editor?.apply()
+                val url = URL("http://jedioleg.asuscomm.com:8000")
+                val connection: HttpURLConnection = url.openConnection() as HttpURLConnection
+                connection.connectTimeout = 500
+                if (connection.responseCode == HttpURLConnection.HTTP_OK) {
+                    try {
+                        val stream = BufferedInputStream(connection.inputStream)
+                        val bufferedReader = BufferedReader(InputStreamReader(stream))
+                        val stringBuilder = StringBuilder()
+                        bufferedReader.forEachLine { stringBuilder.append(it) }
+                        val data: String = stringBuilder.toString()
+                        val json = Json(JsonConfiguration.Stable.copy(isLenient = true, ignoreUnknownKeys = true))
+                        listRates = json.parse(Rate.serializer().list, data)
+                        val db = Room.databaseBuilder(
+                                context,
+                                AppDatabase::class.java, "rates"
+                        ).allowMainThreadQueries().build()
+                        db.rateDao().deleteAll()
+                        db.rateDao().insertAllCountries(listRates!!)
+                        val prefs = context.getSharedPreferences("val", Context.MODE_PRIVATE)
+                        if (prefs?.getString("rate", "") == "") {
+                            val editor = prefs.edit()
+                            editor?.putString("rate", db.rateDao().findByCountry(Locale.getDefault().country).rate.toString())
+                            editor?.apply()
+                        }
+                    }
+                    catch (e: Exception) { e.printStackTrace() }
+                    finally {
+                        connection.disconnect()
+                    }
+                }
+                else {
+                    println("Error ${connection.responseCode}")
                 }
             }
         }
